+++ saleor/extensions/manager.py
@@ -263,16 +263,22 @@ class ExtensionsManager(PaymentInterface):
             )
         raise Exception(f"Payment plugin {gateway} is inaccessible!")
 
-    def list_payment_gateways(self) -> List[dict]:
+    def list_payment_gateways(self, active_only: bool = True) -> List[dict]:
         payment_method = "process_payment"
+        plugins = self.plugins
+        if active_only:
+            plugins = [
+                plugin
+                for plugin in plugins
+                if self.get_plugin_configuration(plugin.PLUGIN_NAME).active
+            ]
         return [
             {
                 "name": plugin.PLUGIN_NAME,
                 "config": self.__get_payment_config(plugin.PLUGIN_NAME),
             }
-            for plugin in self.plugins
+            for plugin in plugins
             if payment_method in type(plugin).__dict__
-            and self.get_plugin_configuration(plugin.PLUGIN_NAME).active
         ]
 
     def get_payment_template(self, gateway: str) -> str: