+++ CHANGELOG.md
@@ -95,6 +95,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - `productVariant` nodes now require `manage_products` permission to query `costPrice` and `stockQuantity` fields. `isAvailable` of a variant is not longer returning false when another variant from the same product is out of stock. - #4753 by @NyanKiyoshi
 - Fixed crash when placing an order when a customer happens to have the same address more than once - #4823 by @NyanKiyoshi
 - Fix fetching staff user without manage_users permission - #4835 by @fowczarek
+- Add form to configure payments in dashboard - #4807 by @szewczykmira
 
 ## 2.8.0